+++ toys-core/lib/toys/standard_mixins/fileutils.rb
@@ -41,6 +41,8 @@ module Toys
     #     include :fileutils
     #
     module Fileutils
+      include Mixin
+
       ## @private
       def self.included(mod)
         mod.include(::FileUtils)